updated block tags to allow custom wood to craft.
updated the block tags for the two types of custom wood so that they can actually turn from logs to planks (including logs, wood, and the stripped variant of both). also updated the branch name of feature/feature_sequoia to feature/feature_custom_wood to better suit the changes being made in this branch.